Add files via upload
Masks_without masks is the project on real time images processing using Deep Learning Algorithms. As the whole world is facing the COVID 19 pandemic , masks is one of the important thing which everyone has to wear in order to stop the spread.  So, I trained the model using the deep learning algorithms CNNmodel, AlexNetmodel and obtained an accuracy of 96 percent. This model can be saved and used with open CV which will detect the live streaming of image and detect whether the person is wearing the mask or not.